pyproject.toml
[project]
name = "ai-agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
